--- test/e2e-test/addon-test/util.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2021 The KubeVela Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"fmt"
-	"os/exec"
-)
-
-func checkAppStatus(addonName string)  {
-	cmd := exec.Command("vela","status", "-n", "vela-system", "addon-" + addonName)
-	fmt.Println(cmd.String())
-	stdout, err := cmd.StdoutPipe()
-	cmd.Stderr = cmd.Stdout
-	if err != nil {
-		panic(err)
-	}
-	if err = cmd.Start(); err != nil {
-		fmt.Println(err)
-	}
-	for {
-		tmp := make([]byte, 1024)
-		_, err := stdout.Read(tmp)
-		fmt.Print(string(tmp))
-		if err != nil {
-			break
-		}
-	}
-	if err = cmd.Wait(); err != nil {
-		fmt.Println(err)
-	}
-}
-
-func checkPodStatus(addonName string) {
-
-}
